Bridge Done

+++ routes/trequest/trequestrecomment_delete.js
@@ -5,16 +5,16 @@ const db = require('../../module/pool.js');
 const moment = require('moment');
 
 router.post('/', async (req, res) => {
-
+	let userIdx = req.body.userIdx;
 	let ircmtIdx = req.body.ircmtIdx;
 
-	if (!ircmtIdx) {
+	if (!ircmtIdx||!userIdx) {
 		res.status(400).send({
 			message : "Null Value"
 		});
 	} else {
-		    let deleteQuery = 'DELETE FROM Irecomment WHERE ircmtIdx=?'
-			let deleteResult = await db.queryParam_Arr(deleteQuery,[ircmtIdx]);
+		    let deleteQuery = 'DELETE FROM Irecomment WHERE ircmtIdx=? and userIdx = ?'
+			let deleteResult = await db.queryParam_Arr(deleteQuery,[ircmtIdx, userIdx]);
 
 			if (!deleteResult) {
 				res.status(500).send({
